+++ packages/prensa/components/Article/Embeds/Facebook/FacebookEmbed.tsx
@@ -1,5 +1,4 @@
 import React from 'react'
-import { FacebookProvider, EmbeddedPost } from 'react-facebook'
 
 import { Container } from '../styled'
 import { FacebookEmbedProps } from './types'
@@ -33,16 +32,6 @@ const FacebookEmbed = ({
     />
   )
 
-  const Web = () => (
-    <FacebookProvider appId={fbappid}>
-      <EmbeddedPost
-        href={url}
-        width={width[1]}
-        {...elementProps}
-      />
-    </FacebookProvider>
-  )
-
   return (
     <Container
       $height={height}
@@ -52,7 +41,7 @@ const FacebookEmbed = ({
       mb={mb}
       ml={ml}
     >
-      {amp ? <Amp /> : <Web />}
+      {amp ? <Amp /> : <></>}
     </Container>
   )
 }