+++ packages/xprog-prensa/news/SideMenu/index.ts
@@ -1,0 +1,2 @@
+export * from './component'
+export { SideMenuContent, SideMenuFooter, SideMenuHeader } from './styles'